Implement dummy sysfs interface
This commit adds a user-facing interface to enable snapshot
management to ouichefs. To this end, a new sysfs entry for each
mounted ouichefs is added into /sys/fs/ouichefs/<device>.

Each partition has four files: create, list, destroy, restore
- Create assigns a unique ID and saves the current state of the
  partition
- List displays all saved snapshots with ID and creation date
- Destroy deletes a snapshot by ID, reclaiming associated data and
  metadata
- Restore restores a snapshot by ID, rolling back the partition to
  its previous state

This interface is not functional right now, it is a dummy.

Reviewed-by: Valentin Geyer <valentin.geyer@rwth-aachen.de>
Co-authored-by: Elena Kremers <elena.kremers@rwth-aachen.de>