Implement reference counting data blocks
This commit lays the groundwork for Copy-on-Write data blocks by providing
generic function to handle allocation and deallocation of data blocks, which
keep track of who still references a block to only release it once it is no
longer referenced.

This implementation uses a new block region, the metadata blocks, to store
a reference counter for each data block. These counters are updated
whenever a new inode or index block references a block (or such a block
stops referencing it).

In addition, a bug was fixed in file truncation.