Implement reference counting data blocks
This commit lays the groundwork for Copy-on-Write data blocks, which will
be used to implement snapshotting.

This implementation uses a new block region, the metadata blocks, to store
a reference counter for each data block. These counters are updated
whenever a new inode or index block references a block (or such a block
stops referencing it).