Implement Snapshots by copying all inodes
This commit implements the 'restore/create/delete' sysfs
operations and thus is the first feature-complete commit
of ouichefs's snapshotting system. This came at a high price:
The inodes are copied, and the space necessary for that is reserved
in mkfs, leading to ~70% of the file system being unusable. This will
be addressed in a later patch.

In addition, the block allocation function were changed to allow
easier implementation of new data block types, copy-on-write was
implemented for directories and some inode functions
were split to allow code reuse. Ideally these would be separat patches.